Removed menu from wrong activity; fixing
